+++ x-pack/plugins/security_solution/server/lib/telemetry/constants.ts
@@ -5,6 +5,8 @@
  * 2.0.
  */
 
+export const TELEMETRY_MAX_BUFFER_SIZE = 100;
+
 export const TELEMETRY_CHANNEL_LISTS = 'security-lists';
 
 export const TELEMETRY_CHANNEL_ENDPOINT_META = 'endpoint-metadata';
