+++ x-pack/plugins/security_solution/server/lib/telemetry/tasks/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { TelemetryDiagTask as DiagnosticTask } from './diagnostic';
+export { TelemetryEndpointTask as EndpointTask } from './endpoint';
+export { TelemetryExceptionListsTask as ExceptionListsTask } from './security_lists';
