+++ x-pack/plugins/security_solution/server/lib/telemetry/tasks/security_lists.test.ts
@@ -6,14 +6,14 @@
  */
 
 import { loggingSystemMock } from 'src/core/server/mocks';
-import { TaskStatus } from '../../../../task_manager/server';
-import { taskManagerMock } from '../../../../task_manager/server/mocks';
-
+import { TaskStatus } from '../../../../../task_manager/server';
+import { taskManagerMock } from '../../../../../task_manager/server/mocks';
+import { TelemetryExceptionListsTask, TelemetrySecuityListsTaskConstants } from './security_lists';
 import {
-  TelemetryExceptionListsTask,
-  TelemetrySecuityListsTaskConstants,
-} from './security_lists_task';
-import { createMockTelemetryEventsSender, MockExceptionListsTask } from './mocks';
+  createMockTelemetryEventsSender,
+  MockExceptionListsTask,
+  createMockTelemetryReceiver,
+} from '../mocks';
 
 describe('test exception list telemetry task functionality', () => {
   let logger: ReturnType<typeof loggingSystemMock.createLogger>;
@@ -26,7 +26,8 @@ describe('test exception list telemetry task functionality', () => {
     const telemetryTrustedAppsTask = new TelemetryExceptionListsTask(
       logger,
       taskManagerMock.createSetup(),
-      createMockTelemetryEventsSender(true)
+      createMockTelemetryEventsSender(true),
+      createMockTelemetryReceiver()
     );
 
     expect(telemetryTrustedAppsTask).toBeInstanceOf(TelemetryExceptionListsTask);
@@ -34,7 +35,12 @@ describe('test exception list telemetry task functionality', () => {
 
   test('the exception list task should be registered', () => {
     const mockTaskManager = taskManagerMock.createSetup();
-    new TelemetryExceptionListsTask(logger, mockTaskManager, createMockTelemetryEventsSender(true));
+    new TelemetryExceptionListsTask(
+      logger,
+      mockTaskManager,
+      createMockTelemetryEventsSender(true),
+      createMockTelemetryReceiver()
+    );
 
     expect(mockTaskManager.registerTaskDefinitions).toHaveBeenCalled();
   });
@@ -44,7 +50,8 @@ describe('test exception list telemetry task functionality', () => {
     const telemetryTrustedAppsTask = new TelemetryExceptionListsTask(
       logger,
       mockTaskManagerSetup,
-      createMockTelemetryEventsSender(true)
+      createMockTelemetryEventsSender(true),
+      createMockTelemetryReceiver()
     );
 
     const mockTaskManagerStart = taskManagerMock.createStart();
@@ -54,8 +61,9 @@ describe('test exception list telemetry task functionality', () => {
 
   test('the exception list task should not query elastic if telemetry is not opted in', async () => {
     const mockSender = createMockTelemetryEventsSender(false);
+    const mockReceiver = createMockTelemetryReceiver();
     const mockTaskManager = taskManagerMock.createSetup();
-    new MockExceptionListsTask(logger, mockTaskManager, mockSender);
+    new MockExceptionListsTask(logger, mockTaskManager, mockSender, mockReceiver);
 
     const mockTaskInstance = {
       id: TelemetrySecuityListsTaskConstants.TYPE,
@@ -76,16 +84,18 @@ describe('test exception list telemetry task functionality', () => {
       ].createTaskRunner;
     const taskRunner = createTaskRunner({ taskInstance: mockTaskInstance });
     await taskRunner.run();
-    expect(mockSender.fetchTrustedApplications).not.toHaveBeenCalled();
+    expect(mockReceiver.fetchTrustedApplications).not.toHaveBeenCalled();
   });
 
   test('the exception list task should query elastic if telemetry opted in', async () => {
     const mockSender = createMockTelemetryEventsSender(true);
     const mockTaskManager = taskManagerMock.createSetup();
+    const mockReceiver = createMockTelemetryReceiver();
     const telemetryTrustedAppsTask = new MockExceptionListsTask(
       logger,
       mockTaskManager,
-      mockSender
+      mockSender,
+      mockReceiver
     );
 
     const mockTaskInstance = {
